Fix AHPlugin build errors
Errors fixed relating the the main plugin returning a list of doubles
instead of list of strings.
Separated the API of returning card values and setting the advice
text. It didn't make much sense to have them together since the
advice technically isn't a card value but a separate part of the
overlay